feat: add CopyButton component for improved code copying functionality in Install page
